% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resids.R
\name{annual}
\alias{annual}
\title{annual composition plot for RTMB model}
\usage{
annual(obs, pred, ind, yrs, label)
}
\arguments{
\item{obs}{observed comp - input sample size and comp weighting adjusted}

\item{pred}{predicted comp}

\item{ind}{vector of ages or lengths}

\item{label}{axis label}
}
\description{
annual composition plot for RTMB model
}
\examples{
\dontrun{
obs = fish_age_obs
pred = report1$fish_age_pred
ind = ages
annual(obs, pred, ind, label = 'Age')
}
}
