% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/francis.R
\name{calculate_francis_weight}
\alias{calculate_francis_weight}
\title{Calculate Francis Reweighting Multiplier (Method TA1.8)}
\usage{
calculate_francis_weight(rpt, data, comp_type = "fish_size")
}
\arguments{
\item{rpt}{The report object from an RTMB model run.}

\item{data}{The data used in the RTMB model run}

\item{comp_type}{A string identifying the composition type, e.g., "fish_age", "srv_age", "fish_size".}
}
\value{
The model report with recommended weight
}
\description{
Calculate Francis Reweighting Multiplier (Method TA1.8)
}
