% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{drop_wt}
\alias{drop_wt}
\title{Drop Data Weight for Sensitivity Analysis}
\usage{
drop_wt(output, item = "srv_wt")
}
\arguments{
\item{output}{A list containing the baseline model output. Must contain the data (`$dat`), 
objective function environment (`$obj`), the compiled model (`$model`), and bounds 
(`$lower`, `$upper`).}

\item{item}{A character string specifying the name of the weighting parameter in the 
`$dat` list to set to zero. Default is `"srv_wt"`.}
}
\value{
A new model run object generated by `run_model()`.
}
\description{
Performs a sensitivity by dropping the weight of a specific 
data component entirely (setting it to zero) and re-running the model.
}
