% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{get_pars}
\alias{get_pars}
\title{pull parameters from report and projection}
\usage{
get_pars(output, model = "Model", addl = NULL, exclude = NULL)
}
\arguments{
\item{output}{The output from the `RTMButils::run_model()` function for the
full dataset. Must contain `obj` (from RTMB::MakeADFun) and `rpt` (the report)}

\item{model}{model name for column}

\item{addl}{any additional parameters to pull}

\item{exclude}{any parameters to exclude}
}
\description{
pull parameters from report and projection
}
