% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jitter_pars}
\alias{jitter_pars}
\title{Jitter parameters}
\usage{
jitter_pars(pars, jitter_amt = 0.1, seed = NULL)
}
\arguments{
\item{pars}{list of named parameters}

\item{jitter_amt}{how much to jitter}

\item{seed}{for reproducibility, default: NULL}
}
\value{
list of named, jittered parameters
}
\description{
Jitter parameters
}
