% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{plot_profile}
\alias{plot_profile}
\title{plot parameter profile}
\usage{
plot_profile(data, exp = FALSE)
}
\arguments{
\item{data}{profile data.frame with columns value, log_like, ...}

\item{exp}{exponentiate the parameter value?}
}
\description{
plot parameter profile
}
