% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_rec_ssb}
\alias{plot_rec_ssb}
\title{Recruitment/SSB plot}
\usage{
plot_rec_ssb(output, folder = NULL, save = TRUE)
}
\arguments{
\item{output}{RTMB::run_model() output}

\item{folder}{folder model is in}

\item{save}{default is TRUE, saves fig to the folder the model is in}

\item{year}{model year}
}
\description{
Recruitment/SSB plot
}
\examples{
plot_rec_ssb(year=2026, output=m24_26, folder="m24-2026")
}
