% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_size_comps}
\alias{plot_size_comps}
\title{Plot size compositions}
\usage{
plot_size_comps(year, output, folder, save = TRUE, type)
}
\arguments{
\item{year}{= assessment year}

\item{output}{RTMButils model run output}

\item{folder}{= folder the model lives in}

\item{save}{default is TRUE, saves fig to the folder the model is in}

\item{type}{'fishery' or 'survey'}
}
\description{
Plot size compositions
}
\examples{
plot_size_comps(year, output, folder, type = "fishery")
}
