% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{profiles}
\alias{profiles}
\title{Profile a parameter using RTMB with re-optimization}
\usage{
profiles(
  output,
  par_name = "log_M",
  par_values = log(seq(0.03, 0.1, 0.005)),
  derived = c("like_fish_age", "like_srv_age", "like_srv", "like_fish_size")
)
}
\arguments{
\item{output}{RTMButils::run_model output}

\item{par_name}{parameter to profile (e.g., "log_M")}

\item{par_values}{values to profile the parameter over}

\item{derived}{character vector of derived quantity likelihoods to extract from report}
}
\value{
A data frame with parameter value, log-likelihood, and derived values
}
\description{
Profile a parameter using RTMB with re-optimization
}
