% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_bio.R
\name{proj_bio}
\alias{proj_bio}
\title{Projected biomass for 2 years}
\usage{
proj_bio(report, Tproj = 2)
}
\arguments{
\item{report}{RTMB report object}

\item{Tproj}{number of years to project, default: 2}
}
\description{
Projected biomass for 2 years
}
