% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resids.R
\name{qq}
\alias{qq}
\title{one-step ahead residuals qq plot for RTMB model}
\usage{
qq(obs, pred, iss, yrs, label)
}
\arguments{
\item{obs}{observed comp}

\item{pred}{predicted comp}

\item{iss}{input sample size}

\item{yrs}{comp years}

\item{label}{axis label}
}
\description{
one-step ahead residuals qq plot for RTMB model
}
\examples{
\dontrun{
iss = data$fish_age_iss
obs = fish_age_obs
pred = report1$fish_age_pred
yrs = fish_age_yrs
label = "Age"
osa(obs = fish_age_obs, pred = report1$fish_age_pred, iss = data$fish_age_iss,
    yrs = fish_age_yrs, ind = ages) +  ggtitle('osa fishery age comp residuals')
}
}
