% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resids.R
\name{resids}
\alias{resids}
\title{get all residual plots for RTMB model}
\usage{
resids(object, var = "fish_age", outlier = 3, label = "Age", addCI = TRUE)
}
\arguments{
\item{object}{RTMButils::run_model() output}

\item{var}{variable name, default "fish_age"}

\item{outlier}{outlier size, default: 3}

\item{label}{axis label}

\item{addCI}{add CI to qq plot}
}
\description{
get all residual plots for RTMB model
}
\examples{
\dontrun{
fish_age_resids <- resids(output = model_1)
fish_age_resids$osa + ggtitle('osa fishery age comp residuals')
}
}
