% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro.R
\name{run_prospective}
\alias{run_prospective}
\title{Run a prospective analysis (peeling from the start) for an RTMB stock assessment model.}
\usage{
run_prospective(
  output,
  n_peels = 5,
  year,
  folder,
  subfolder = NULL,
  quantities = c("spawn_bio", "tot_bio", "recruits"),
  peel_pars = c("log_Ft", "log_Rt"),
  save_outputs = TRUE
)
}
\arguments{
\item{output}{The output list from the `run_model()` function.}

\item{n_peels}{The number of historical years to remove from the start (integer).}

\item{year}{The assessment year (for file paths).}

\item{folder}{The folder name (for file paths).}

\item{subfolder}{Optional subfolder (for file paths).}

\item{quantities}{Variable names to analyze (e.g., c("spawn_bio", "recruits")).}

\item{peel_pars}{names of parameters with annual index}

\item{save_outputs}{Logical. If TRUE, saves results and plots.}
}
\description{
Run a prospective analysis (peeling from the start) for an RTMB stock assessment model.
}
