% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectivity.R
\name{sel_gamma}
\alias{sel_gamma}
\title{MESA gamma selectivity}
\usage{
sel_gamma(age, b50, delta, adj = 0)
}
\arguments{
\item{age}{age to examine}

\item{b50}{peak point}

\item{delta}{steepness of slope}

\item{adj}{based upon index that starts at 1 - unless adjusted to start at recruitment age
doesn't really do anything other than scale}
}
\description{
MESA gamma selectivity
}
