% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectivity.R
\name{sel_logistic}
\alias{sel_logistic}
\title{standard MESA logistic selectivity}
\usage{
sel_logistic(age, a50, delta, adj)
}
\arguments{
\item{age}{age to examine}

\item{a50}{inflection point}

\item{delta}{steepness of slope}

\item{adj}{based upon index that starts at 1 - unless adjusted to start at recruitment age
doesn't really do anything other than scale}
}
\description{
standard MESA logistic selectivity
}
