% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{solve_Fx}
\alias{solve_Fx}
\title{Solve for target F
differentiable bisection solver}
\usage{
solve_Fx(target, M, slx, wt_mat, sp_fract, A, SB0_val)
}
\arguments{
\item{target}{the target F ratio}

\item{M}{natural mortality}

\item{slx}{fishery selectivity (dims: A,T)}

\item{wt_mat}{mature weight at age}

\item{sp_fract}{when spawning occurs}

\item{A}{number of age classes}

\item{SB0_val}{spawning biomass per recruit}
}
\description{
Solve for target F
differentiable bisection solver
}
