% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resids.R
\name{agg}
\alias{agg}
\title{aggregate residual plot for RTMB model}
\usage{
agg(obs, pred, ind, label)
}
\arguments{
\item{obs}{observed comp - input sample size and comp weighting adjusted}

\item{pred}{predicted comp}

\item{ind}{vector of ages or lengths}

\item{label}{axis label}
}
\description{
aggregate residual plot for RTMB model
}
\examples{
\dontrun{
obs = fish_age_obs
pred = report1$fish_age_pred
ind = ages
agg(obs, pred, ind, label = 'Age')
}
}
