% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{get_likes}
\alias{get_likes}
\title{pull likelihoods from report}
\usage{
get_likes(rpt, model = "Model", addl = NULL, exclude = NULL)
}
\arguments{
\item{rpt}{The output list from the `run_model()` function for the
full dataset. Must contain `obj` (from RTMB::MakeADFun) and `rpt` (the report)}

\item{model}{name for column}

\item{addl}{any additional parameters to pull}

\item{exclude}{any parameters to exclude}
}
\description{
pull likelihoods from report
}
