% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resids.R
\name{osa}
\alias{osa}
\title{one-step ahead residuals for RTMB model}
\usage{
osa(obs, pred, yrs, ind, label = "Age", outlier = 3, addCI = TRUE)
}
\arguments{
\item{obs}{observed comp - input sample size and comp weighting adjusted}

\item{pred}{predicted comp}

\item{yrs}{comp years}

\item{ind}{vector or ages of lengths}

\item{label}{axis label}

\item{outlier}{mark point as outlier if greater than. default: 3}

\item{addCI}{boolean to include 95% confidence intervals on the SDNR label, default:TRUE}
}
\description{
one-step ahead residuals for RTMB model
}
\examples{
\dontrun{
obs = fish_age_obs
pred = report1$fish_age_pred
yrs = fish_age_yrs
label = "Age"
ind = ages
osa(obs = fish_age_obs, pred = report1$fish_age_pred, iss = data$fish_age_iss,
    yrs = fish_age_yrs, ind = ages) +  ggtitle('osa fishery age comp residuals')
}
}
