% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resids.R
\name{pearson}
\alias{pearson}
\title{pearson residuals for RTMB model}
\usage{
pearson(obs, pred, iss, yrs, wt, ind, outlier, label, theta = NULL)
}
\arguments{
\item{obs}{observed comp - input sample size and comp weighting adjusted}

\item{pred}{predictedcomp}

\item{iss}{input sample size}

\item{yrs}{comp years}

\item{wt}{comp data weighting}

\item{ind}{vector or ages of lengths}

\item{label}{axis label}

\item{theta}{optional scalar for Dirichlet-multinomial}
}
\description{
pearson residuals for RTMB model
}
\examples{
\dontrun{
pearson(obs = fish_age_obs * iss * wt, pred = report1$fish_age_pred, iss = data$fish_age_iss,
    yrs = fish_age_yrs, ind = ages) +  ggtitle('osa fishery age comp residuals')
}
}
