% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_phase_plane}
\alias{plot_phase_plane}
\title{Phase plane plot}
\usage{
plot_phase_plane(..., year, output, folder = NULL, save = TRUE, shade = FALSE)
}
\arguments{
\item{...}{add'l ggplot inputs}

\item{year}{model year}

\item{output}{RTMB::run_model() output}

\item{folder}{folder model is in}

\item{save}{default is TRUE, saves fig to the folder the model is in}
}
\description{
Phase plane plot
}
\examples{
plot_phase_plane(year=2026, output=m24_26, folder="m24-2026")
}
