% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resids.R
\name{resids}
\alias{resids}
\title{get all residual plots for RTMB model}
\usage{
resids(obs, pred, iss, outlier = 3, yrs, ind, label = "Age")
}
\arguments{
\item{obs}{observed comp}

\item{pred}{predictedcomp}

\item{iss}{input sample size}

\item{outlier}{outlier size, default: 3}

\item{yrs}{comp years}

\item{ind}{vector or ages of lengths}

\item{label}{axis label}
}
\description{
get all residual plots for RTMB model
}
\examples{
\dontrun{
iss = data$fish_age_iss
obs = fish_age_obs
pred = report1$fish_age_pred
yrs = fish_age_yrs
label = "Age"
ind = ages
fish_age_resids <- resids(obs = fish_age_obs, pred = report1$fish_age_pred, iss = data$fish_age_iss,
                         yrs = fish_age_yrs, ind = ages)
 fish_age_resids$osa + ggtitle('osa fishery age comp residuals')
}
}
