% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_model}
\alias{run_model}
\title{Run RTMB model}
\usage{
run_model(
  model,
  data,
  pars,
  map = NULL,
  proj = TRUE,
  lower = NULL,
  upper = NULL,
  random = NULL,
  newton_loops = 3,
  control = list(iter.max = 1e+05, eval.max = 20000, rel.tol = 1e-12)
)
}
\arguments{
\item{model}{the model function}

\item{data}{named data list}

\item{pars}{named parameter list}

\item{map}{named mapping list, default: NULL}

\item{proj}{whether to run the projection module: TRUE}

\item{lower}{unnamed vector of lower parameter limits, default: NULL}

\item{upper}{unnamed vector of upper parameter limits, default: NULL}

\item{random}{vector of parameter(s) to be random effects, default: NULL}

\item{newton_loops}{number of newton loops to run to reduce gradient: 3 - note only works for unconstrained models}

\item{control}{optimization settings list(iter.max = 1e5, eval.max = 2e4, rel.tol = 1e-12)}
}
\description{
Run RTMB model
}
