% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_model}
\alias{run_model}
\title{Run RTMB model}
\usage{
run_model(
  model,
  data,
  pars,
  map = NULL,
  lower = NULL,
  upper = NULL,
  random = NULL
)
}
\arguments{
\item{model}{the model function}

\item{data}{named data list}

\item{pars}{named parameter list}

\item{map}{named mapping list, default: NULL}

\item{lower}{unnamed vector of lower parameter limits, default: NULL}

\item{upper}{unnamed vector of upper parameter limits, default: NULL}

\item{random}{vector of parameter(s) to be random effects, default: NULL}
}
\description{
Run RTMB model
}
