% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/francis.R
\name{run_model_reweight}
\alias{run_model_reweight}
\title{Run model with iterative Francis reweighting for composition data}
\usage{
run_model_reweight(output, iters = 10, ...)
}
\arguments{
\item{output}{The output from RTMButils::run_model() that contains the model, named data list, and named parameter list.}

\item{iters}{number of reweighting iterations.}
}
\description{
Run model with iterative Francis reweighting for composition data
}
