% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/francis.R
\name{run_model_reweight}
\alias{run_model_reweight}
\title{Run model with iterative Francis reweighting for composition data}
\usage{
run_model_reweight(model, data, pars, iters = 10, threshold = 0.05, ...)
}
\arguments{
\item{model}{The model function.}

\item{data}{The initial named data list.}

\item{pars}{The initial named parameter list.}

\item{iters}{number of reweighting iterations.}

\item{...}{Other arguments to pass to run_model (map, lower, upper, random).}
}
\description{
Run model with iterative Francis reweighting for composition data
}
