% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro.R
\name{run_prospective}
\alias{run_prospective}
\title{Run a prospective analysis (peeling) for an RTMB stock assessment model.}
\usage{
run_prospective(
  rpt,
  data,
  pars,
  model,
  n_peels = 10,
  year,
  folder,
  subfolder = NULL,
  quantities = c("spawn_bio", "tot_bio", "recruits"),
  map = list(sigmaR = factor(NA)),
  control = list(iter.max = 1e+05, eval.max = 20000),
  save_outputs = TRUE
)
}
\arguments{
\item{rpt}{The output list from the `run_model()` function for the
full dataset. Must contain `obj` (from RTMB::MakeADFun) and `rpt` (the report).}

\item{data}{A list of data objects used for the full model fit.}

\item{pars}{A list of initial parameter values for the full model fit.}

\item{model}{The R code object representing the model structure to be passed to RTMB::MakeADFun.}

\item{n_peels}{The number of historical years to remove for the analysis (integer).}

\item{year}{The assessment year, used for constructing file paths (e.g., 2024).}

\item{folder}{The folder name for the model or type (e.g., "mgmt", "research"), used for file paths.}

\item{subfolder}{An optional subfolder name for file paths (e.g., "m24"), default: NULL.}

\item{quantities}{A character vector of variable names from the model report
to analyze (e.g., c("spawn_bio", "tot_bio", "recruits")).}

\item{map}{A named list specifying the mapping for `RTMB::MakeADFun`.}

\item{control}{A list of control parameters for the `nlminb` optimizer.}

\item{save_outputs}{A logical flag. If TRUE, saves the prospective results
(.RDS) and plots (.png) to a 'retro' subfolder.}
}
\value{
A list containing two elements:
  1. `prospect_data`: A tidy data frame with all estimates and standard errors from all peels.
  2. `plots`: A named list of the generated ggplot objects.
}
\description{
This function automates the process of prospective analysis. It iteratively
removes historical years of data, refits the model, and generates diagnostic
plots for specified quantities of interest. It is a complementary diagnostic
to a retrospective analysis, providing insight into the model's sensitivity to
historical data.
}
\details{
The function relies on the following packages: `RTMB`, `dplyr`, `purrr`,
`ggplot2`, `patchwork`, `scico`, `tibble`, and `here`. Ensure they are installed.
In contrast to a retrospective analysis, a prospective analysis "peels" off historical data
to see how the model's estimates for key quantities evolve and stabilize with more data.
It is a useful diagnostic for understanding how much the model's conclusions rely on early
data points.
}
