% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resids.R
\name{sample_size}
\alias{sample_size}
\title{Composition sample size}
\usage{
sample_size(obs, pred, iss, yrs, wt, theta = NULL)
}
\arguments{
\item{obs}{observed comp data- input sample size and comp weighting adjusted}

\item{pred}{predicted comp data}

\item{iss}{input sample size}

\item{yrs}{years}

\item{wt}{comp weighting}

\item{theta}{optional scalar for Dirichlet-multinomial}
}
\description{
Composition sample size
}
\examples{
\dontrun{
sample_size(obs, pred, iss, yrs)}
}
